+++ kinectMocap.h
@@ -31,6 +31,7 @@ using namespace boost::python;
 
 // Current Kinect
 IKinectSensor*          m_pKinectSensor;
+double					tilt, sensorH, planeX, planeY, planeZ;
 
 Joint					joints[JointType_Count];